Porting over changes that allow incremental typescript updates using swc and related tools
