+++ .vscode/launch.json
@@ -2,17 +2,12 @@
   "version": "0.2.0",
   "configurations": [
     {
-      "env": {
-        "NODE_ENV": "development"
-      },
-      "name": "Debug Server",
-      "outputCapture": "std",
-      "program": "${workspaceFolder}/src/server.js",
-      "request": "launch",
-      "skipFiles": [
-        "<node_internals>/**"
-      ],
-      "type": "node"
+      "type": "node",
+      "request": "attach",
+      "name": "Node: Nodemon",
+      "processId": "${command:PickProcess}",
+      "restart": true,
+      "protocol": "inspector"
     }
   ]
 }
