+++ package.json
@@ -13,17 +13,23 @@
     "yarn": ">=1.5.1"
   },
   "scripts": {
-    "start": "nodemon ./src/server.js --exec 'npm run lint && node --inspect' | bunyan",
+    "clean": "rm -rf dist",
+    "start": "nodemon ./src/server.ts --exec 'npm run lint && node -r @swc-node/register' | bunyan",
     "start:dev": "export NODE_ENV=development && npm run start",
-    "test": "NODE_ENV=test nodemon ./src/test.js",
-    "test:ci": "NODE_ENV=test nyc --reporter=lcov --reporter=text-summary node ./src/test.js --ci",
-    "test:gitlab": "NODE_ENV=testgitlab nyc --reporter=lcov --reporter=text-summary node ./src/test.js --ci",
-    "test:debug": "NODE_ENV=test nodemon ./src/test.js --exec 'node --inspect-brk=:9339'",
+    "start:prod": "cd dist && node ./src/server.js",
+    "test": "NODE_ENV=test nodemon -r @swc-node/register ./src/test.js",
+    "test:build": "cd dist && node -r @swc-node/register ./src/test.js",
+    "test:ci": "NODE_ENV=test nyc --reporter=lcov --reporter=text-summary node -r @swc-node/register ./src/test.js --ci",
+    "test:gitlab": "NODE_ENV=testgitlab nyc --reporter=lcov --reporter=text-summary node -r @swc-node/register ./src/test.js --ci",
+    "test:debug": "NODE_ENV=test nodemon ./src/test.js --exec 'node -r @swc-node/register --inspect-brk=:9339'",
     "build": "npm run lint",
+    "build:prod": "npm run lint && npm run clean && npm run compile:prod && npm run link:prod",
+    "compile:prod": "swc src --config-file .prod.swcrc -d dist/src --copy-files",
+    "link:prod": "ln -s ../config ./dist/config && ln -s ../node_modules ./dist/node_modules && ln -s ../package.json ./dist/package.json && ln -s ../package-lock.json ./dist/package-lock.json",
     "lint": "npm run lint:eslint",
     "lint:fix": "npm run lint:eslint:fix",
-    "lint:eslint": "eslint \"./**/*.js\"",
-    "lint:eslint:fix": "eslint \"./**/*.js\" --fix",
+    "lint:eslint": "eslint \"./**/*.js\" \"./**/*.ts\"",
+    "lint:eslint:fix": "eslint \"./**/*.js\" \"./**/*.ts\" --fix",
     "tsc:check": "tsc --project tsconfig.json --noEmit",
     "tsc:watch": "tsc --watch --noEmit",
     "prepare": "husky install"
@@ -79,6 +85,9 @@
     "yargs": "^17.5.1"
   },
   "devDependencies": {
+    "@swc-node/register": "^1.5.1",
+    "@swc/cli": "^0.1.57",
+    "@swc/core": "^1.2.233",
     "@types/express": "4.16",
     "@types/lodash": "^4.14.176",
     "@types/luxon": "^1.27.1",
@@ -86,6 +95,8 @@
     "@types/passport": "0.4",
     "@types/q": "1.5",
     "@types/yargs": "^17.0.10",
+    "@typescript-eslint/eslint-plugin": "^5.33.1",
+    "@typescript-eslint/parser": "^5.33.1",
     "eslint": "^7.31.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-jsdoc": "^35.5.0",
